refactor posts filter
